ORAD group/requirements.txt
alembic==1.12.0
asgiref==3.7.2
asttokens==2.4.1
decorator==5.1.1
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
executing==2.0.1
greenlet==2.0.2
ipdb==0.13.13
ipython==8.22.1
jedi==0.19.1
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
stack-data==0.6.3
tabulate==0.9.0
traitlets==5.14.1
typing_extensions==4.9.0
wcwidth==0.2.13
